SecureNestt/backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
python-jose==3.5.0
email-validator
passlib[bcrypt]
python-multipart
cohere
AppOpener
beautifulsoup4
groq
rich
googlesearch-python
